# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bifrost",
    "version": "1.4.0",
    "description": "The HTTP API of a Bifrost LAN hub — the same one its web app and its `bifrost` CLI use.\n\nTwo things to know when reading this description:\n\n- **Sibling path parameters.** Each document kind reads a record at `/api/<kind>/{slug}` and\n  writes it at `/api/<kind>/{id}`. OpenAPI treats those as one path with two parameter\n  names, which it forbids; renaming a route parameter is a code change made only for the\n  description's sake, so both appear as written and the merge is left to the API docs plan.\n- **Request validation follows Fastify's defaults.** Properties a request schema does not\n  declare are stripped rather than refused, and scalar values are coerced to the declared\n  type (`\"7\"` for an integer is accepted as `7`)."
  },
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bifrost_admin",
        "description": "The Heimdall session cookie set by POST /api/heimdall/login"
      }
    },
...
  },
  "paths": {
...
    "/api/atlas": {
      "get": {
        "operationId": "listAtlasDocs",
        "summary": "List saved XML documents",
        "tags": [
          "atlas"
        ],
        "description": "Filter by name (`q`) or author device (`author`) and sort. `paged=true` opts into the offset envelope (PLAN-31); without it the response is the legacy bare array.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 120
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "in": "query",
            "name": "author",
            "required": false
          },
          {
            "schema": {
              "enum": [
                "name",
                "created",
                "modified",
                "size"
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "enum": [
                "asc",
                "desc"
              ]
            },
            "in": "query",
            "name": "order",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "paged",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Without `paged=true`, the bare array it always was; with it, one page",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Without `paged=true`, the bare array it always was; with it, one page",
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "slug",
                          "authorDeviceId",
                          "sizeBytes",
                          "createdAt",
                          "modifiedAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string",
                            "description": "`<kebab-name>-<id>`; changes on rename, and a stale slug answers 301"
                          },
                          "authorDeviceId": {
                            "type": [
                              "null",
                              "string"
                            ],
                            "description": "The saving device, or null when it was not known"
                          },
                          "sizeBytes": {
                            "type": "integer",
                            "description": "UTF-8 bytes of the content"
                          },
                          "createdAt": {
                            "type": "integer",
                            "description": "Unix epoch milliseconds"
                          },
                          "modifiedAt": {
                            "type": "integer",
                            "description": "Unix epoch milliseconds"
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "items",
                        "total",
                        "limit",
                        "offset",
                        "authors"
                      ],
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "slug",
                              "authorDeviceId",
                              "sizeBytes",
                              "createdAt",
                              "modifiedAt"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string",
                                "description": "`<kebab-name>-<id>`; changes on rename, and a stale slug answers 301"
                              },
                              "authorDeviceId": {
                                "type": [
                                  "null",
                                  "string"
                                ],
                                "description": "The saving device, or null when it was not known"
                              },
                              "sizeBytes": {
                                "type": "integer",
                                "description": "UTF-8 bytes of the content"
                              },
                              "createdAt": {
                                "type": "integer",
                                "description": "Unix epoch milliseconds"
                              },
                              "modifiedAt": {
                                "type": "integer",
                                "description": "Unix epoch milliseconds"
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer",
                          "description": "Rows matching the filters, across every page"
                        },
                        "limit": {
                          "type": "integer",
                          "description": "The page size the server used"
                        },
                        "offset": {
                          "type": "integer"
                        },
                        "authors": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "Every author device id over the kind's whole table, unfiltered"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The request failed validation",
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable code, e.g. NOT_FOUND"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAtlasDoc",
        "summary": "Save a new XML document",
        "tags": [
          "atlas"
        ],
        "description": "An omitted or blank name gets a generated, collision-free one.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "authorDeviceId",
                    "sizeBytes",
                    "createdAt",
                    "modifiedAt",
                    "content"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string",
                      "description": "`<kebab-name>-<id>`; changes on rename, and a stale slug answers 301"
                    },
                    "content": {
                      "type": "string",
                      "description": "The document text exactly as saved"
                    },
                    "authorDeviceId": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "description": "The saving device, or null when it was not known"
                    },
                    "sizeBytes": {
                      "type": "integer",
                      "description": "UTF-8 bytes of the content"
                    },
                    "createdAt": {
                      "type": "integer",
                      "description": "Unix epoch milliseconds"
                    },
                    "modifiedAt": {
                      "type": "integer",
                      "description": "Unix epoch milliseconds"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The request failed validation",
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable code, e.g. NOT_FOUND"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Over the size limit",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Over the size limit",
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable code, e.g. NOT_FOUND"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Unsupported media type",
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable code, e.g. NOT_FOUND"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/atlas/api/{slug}": {
      "get": {
        "operationId": "getAtlasDocRaw",
        "summary": "The raw document text, as a public data URL",
        "tags": [
          "atlas"
        ],
        "description": "Outside `/api/` on purpose, with CORS open: a saved document doubles as a stable URL for other tools. `?download` adds an attachment `content-disposition`.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "download",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The stored text, byte for byte",
            "headers": {
              "access-control-allow-origin": {
                "schema": {
                  "type": "string"
                },
                "description": "Always `*`"
              },
              "content-disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "With `?download`: `attachment; filename=\"<name>.xml\"`"
              }
            },
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "301": {
            "description": "A stale slug: the canonical raw URL",
            "headers": {
              "location": {
                "schema": {
                  "type": "string"
                },
                "description": "Where the resource now lives"
              },
              "access-control-allow-origin": {
                "schema": {
                  "type": "string"
                },
                "description": "Always `*`"
              }
            }
          },
          "400": {
            "description": "The request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The request failed validation",
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable code, e.g. NOT_FOUND"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found",
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable code, e.g. NOT_FOUND"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}